move storybook dirs in to src
