--- .storybook/main.ts
@@ -1,21 +1,0 @@
-import type { StorybookConfig } from '@storybook/react-vite'
-
-const config: StorybookConfig = {
-  stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
-
-  addons: [
-    '@storybook/addon-a11y',
-    '@storybook/addon-docs',
-    '@storybook/addon-vitest',
-    '@chromatic-com/storybook'
-  ],
-
-  framework: {
-    name: '@storybook/react-vite',
-    options: {}
-  },
-
-  staticDirs: ['../public']
-}
-
-export default config
