+++ src/storybook/config/preview.ts
@@ -1,9 +1,9 @@
-import '../src/index.css'
+import '@/index.css'
 
 import type { Preview } from '@storybook/react-vite'
 import { initialize, mswLoader } from 'msw-storybook-addon'
 
-import { withTheme } from '../src/stories/decorators'
+import { withTheme } from '../decorators'
 import { globalTypes } from './globalTypes'
 
 // Initialize MSW
