+++ src/storybook/config/theme.ts
@@ -1,0 +1,20 @@
+import { create } from 'storybook/theming'
+
+export default create({
+  base: 'light',
+  // Typography
+  fontBase: '"Open Sans", sans-serif',
+  fontCode: 'monospace',
+
+  // Colours
+  colorPrimary: '#3A10E5',
+  colorSecondary: '#585C6D',
+
+  // UI
+  appBg: '#ffffff',
+  appContentBg: '#ffffff',
+
+  // Text colors
+  textColor: '#10162F',
+  textInverseColor: '#ffffff'
+})
