+++ src/storybook/docs/Introduction.mdx
@@ -1,12 +1,12 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import Code from './assets/code-brackets.svg';
-import Colors from './assets/colors.svg';
-import Comments from './assets/comments.svg';
-import Direction from './assets/direction.svg';
-import Flow from './assets/flow.svg';
-import Plugin from './assets/plugin.svg';
-import Repo from './assets/repo.svg';
-import StackAlt from './assets/stackalt.svg';
+import { Meta } from '@storybook/addon-docs/blocks'
+import Code from '@/assets/code-brackets.svg'
+import Colors from '@/assets/colors.svg'
+import Comments from '@/assets/comments.svg'
+import Direction from '@/assets/direction.svg'
+import Flow from '@/assets/flow.svg'
+import Plugin from '@/assets/plugin.svg'
+import Repo from '@/assets/repo.svg'
+import StackAlt from '@/assets/stackalt.svg'
 
 <Meta title="Introduction" />
 
@@ -128,22 +128,14 @@ We recommend building UIs with a [**component-driven**](https://componentdriven.
 <div className="subheading">Configure</div>
 
 <div className="link-list">
-  <a
-    className="link-item"
-    href="https://storybook.js.org/docs/react/addons/addon-types"
-    target="_blank"
-  >
+  <a className="link-item" href="https://storybook.js.org/docs/react/addons/addon-types" target="_blank">
     <img src={Plugin} alt="plugin" />
     <span>
       <strong>Presets for popular tools</strong>
       Easy setup for TypeScript, SCSS and more.
     </span>
   </a>
-  <a
-    className="link-item"
-    href="https://storybook.js.org/docs/react/configure/webpack"
-    target="_blank"
-  >
+  <a className="link-item" href="https://storybook.js.org/docs/react/configure/webpack" target="_blank">
     <img src={StackAlt} alt="Build" />
     <span>
       <strong>Build configuration</strong>
@@ -208,6 +200,5 @@ We recommend building UIs with a [**component-driven**](https://componentdriven.
 </div>
 
 <div className="tip-wrapper">
-  <span className="tip">Tip</span>Edit the Markdown in{' '}
-  <code>stories/Introduction.stories.mdx</code>
+  <span className="tip">Tip</span>Edit the Markdown in <code>storybook/docs/Introduction.stories.mdx</code>
 </div>
